requirements.txt
aniso8601==7.0.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.0.3
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
marshmallow==2.19.5
pkg-resources==0.0.0
psycopg2==2.8.3
PyJWT==1.7.1
pyodbc==4.0.26
pytz==2019.1
PyYAML==5.1.1
redis==3.2.1
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.5
urllib3==1.25.3
uWSGI==2.0.18
Werkzeug==0.15.4
